Share closure cells with every stored callback and gate the racer countdown
A stored closure owns its captures by value, so a binding a stored
callback writes must be a cell every closure shares. The analysis knew
listeners, record methods, returned functions and inline arrows handed to
a same-file helper; the racer's countdown re-armed its tick through a
timer and flipped countdownActive into a private copy, so the car took no
input. A local function is now stored wherever it is referenced but as a
direct callee, record accessors are stored members, data-sink closures and
callbacks registered inside another callback are roots, and a retaining
helper may live in any repository module. Sandblox's hinge drivers, quake's
refire gate and freeciv's scout destination regenerate as shared cells.

trace_dynamic_frame takes the frame delta as double so MSVC builds both
PALs warning-clean again.

A project-owned gate observes a timer-flipped and an interval-counted flag
at a fixed frame on both backends; generated with the pre-fix compiler it
fails. The /simplify record covers the change.

+++ native/src/pal_runtime_trace.hpp
@@ -130,7 +130,7 @@ inline void trace_scene_topology(
  */
 inline void trace_dynamic_frame(
     const Engine& engine,
-    float delta_ms,
+    double delta_ms,
     long frame) {
     if (!runtime_trace_enabled()) return;
     if (
